requirements.txt
geopy==2.2.0
timezonefinder==4.5.0
requests==2.28.2
pytz==2023.3
Pillow==9.5.0
